Refactor typescript extension and implement multi-ide gradle support
