+++ sdk/compiler/damlc/lib/DA/Cli/Damlc.hs
@@ -338,6 +338,7 @@ cmdMultiIde _numProcessors =
   where
     cmd = fmap (Command MultiIde Nothing) $ runMultiIde
         <$> cliOptLogLevel
+        <*> strOptionOnce (long "ide-identifier" <> help "Identifier string for this IDE")
         <*> many (strArgument mempty)
 
 cmdIde :: SdkVersion.Class.SdkVersioned => Int -> Mod CommandFields Command
