+++ sdk/compiler/damlc/lib/DA/Cli/Damlc/Command/MultiIde/Prefixing.hs
@@ -11,12 +11,15 @@ module DA.Cli.Damlc.Command.MultiIde.Prefixing (
   addProgressTokenPrefixToClientMessage,
   addProgressTokenPrefixToServerMessage,
   addLspPrefixToServerMessage,
+  addCommandPrefixes,
   stripLspPrefix,
   stripWorkDoneProgressCancelTokenPrefix,
+  stripCommandPrefixes,
 ) where
 
 import Control.Lens
 import Control.Monad
+import Data.Maybe (fromMaybe)
 import qualified Data.Text as T
 import Data.Tuple (swap)
 import qualified Language.LSP.Types as LSP
@@ -210,3 +213,25 @@ addLspPrefixToServerMessage ide res@(LSP.FromServerMess method params) =
       case params of
         LSP.ReqMess params' -> LSP.FromServerMess method $ LSP.ReqMess $ params' & LSP.id %~ addLspPrefix (ideMessageIdPrefix ide)
         LSP.NotMess _ -> res
+
+-- Gradle support requires we allow multiple multi-ides to run at once. This means the commands registered by the IDE must be unique
+
+commandsToPrefix :: [T.Text]
+commandsToPrefix = ["typesignature.add"]
+
+prefixCommand :: MultiIdeState -> T.Text -> T.Text
+prefixCommand miState cmd = if cmd `elem` commandsToPrefix then misIdentifier miState <> "." <> cmd else cmd
+
+-- Adds the unique identifier to the front of commands in code lens replies
+addCommandPrefixes :: MultiIdeState -> LSP.FromServerMessage -> LSP.FromServerMessage
+addCommandPrefixes miState (LSP.FromServerRsp LSP.STextDocumentCodeLens rsp) =
+  let prefixRsp = LSP.result . traverse . traverse . LSP.command . traverse . LSP.command %~ prefixCommand miState
+   in LSP.FromServerRsp LSP.STextDocumentCodeLens $ prefixRsp rsp
+addCommandPrefixes _ msg = msg
+
+-- This function strips the unique identifier we prefix commands with so they can be forwarded to subIDEs as normal
+stripCommandPrefixes :: MultiIdeState -> LSP.FromClientMessage' a -> LSP.FromClientMessage' a
+stripCommandPrefixes miState (LSP.FromClientMess LSP.SWorkspaceExecuteCommand req) =
+  let stripPrefix = LSP.params . LSP.command %~ (\s -> fromMaybe s $ T.stripPrefix (misIdentifier miState <> ".") s)
+   in LSP.FromClientMess LSP.SWorkspaceExecuteCommand $ stripPrefix req
+stripCommandPrefixes _ msg = msg
